requirements.txt
Flask
Flask-SQLAlchemy
PyMySQL
Faker
cryptography
